requirements.txt
altgraph==0.17.2
certifi==2021.10.8
charset-normalizer==2.0.8
future==0.18.2
idna==3.3
numpy==1.21.4
opencv-python==4.5.4.60
pdf2image==1.16.0
pefile==2021.9.3
Pillow==8.4.0
points==0.4.0
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pywin32-ctypes==0.2.0
requests==2.26.0
urllib3==1.26.7
